+++ tests/testthat/to-img.R
@@ -1,3 +1,9 @@
+get_output_file <- function(rmd){
+  basen <- tools::file_path_sans_ext(basename(rmd))
+  list.files(path = dirname(rmd), pattern = paste0(basen, ".(html|pdf|docx|pptx)$"),
+             full.names = TRUE)
+}
+
 compare_image <- function(img1, img2){
   require("magick")
   img.reference <- image_raster(magick::image_read(img1), tidy = FALSE)